+++ straxen/analyses/__init__.py
@@ -2,3 +2,4 @@ from . import quick_checks
 from . import records_matrix
 from . import waveform_plot
 from . import holoviews_waveform_display
+from . import event_display
